update_responses: generate incremental MAR files
Generate incremental MAR file when they are not already there.

The list of versions from which to generate incremental MAR is defined in
config.yml as incremental_from for each version.

The script expects the MAR tools (the mar and make_incremental_update.sh
commands) to be available in the PATH.