update_responses: generate multiple MAR files in parallel
Generating incremental MAR files takes a lot of time but only uses one
thread. To make that faster, we will be creating multiple MAR files in
parallel.

The number of files we generate in parallel will be equal to the number
of CPU cores, unless the NUM_PROCS environment variable is set.